4819 refactoring BP modals
