+++ src/options/pages/blueprints/BlueprintsPage.tsx
@@ -15,52 +15,30 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { useMemo, useEffect } from "react";
+import React, { useEffect } from "react";
 import BlueprintsPageLayout from "@/options/pages/blueprints/BlueprintsPageLayout";
 import useInstallables from "@/options/pages/blueprints/useInstallables";
-import ExtensionLogsModal from "@/options/pages/blueprints/modals/ExtensionLogsModal";
-import { useSelector } from "react-redux";
-import { selectModalsContext } from "@/options/pages/blueprints/modals/blueprintModalsSelectors";
 import { useTitle } from "@/hooks/title";
 import { ErrorDisplay } from "@/layout/ErrorDisplay";
-import ConvertToRecipeModal from "./modals/ConvertToRecipeModal";
-import ShareRecipeModal from "./modals/ShareRecipeModal/ShareRecipeModal";
 import { reportEvent } from "@/telemetry/events";
-import PublishRecipeModal from "./modals/ShareRecipeModal/PublishRecipeModal";
+import Modals from "./modals/Modals";
 
 const BlueprintsPage: React.FunctionComponent = () => {
   useTitle("Blueprints");
   const { installables, error } = useInstallables();
 
-  const { showLogsContext, showShareContext, showPublishContext } =
-    useSelector(selectModalsContext);
-
   useEffect(() => {
     reportEvent("BlueprintsPageView");
   }, []);
 
-  const body = useMemo(() => {
-    if (error) {
-      return <ErrorDisplay error={error} />;
-    }
-
-    return <BlueprintsPageLayout installables={installables} />;
-  }, [installables, error]);
-
   return (
     <div className="h-100">
-      {showLogsContext && (
-        <ExtensionLogsModal
-          title={showLogsContext.title}
-          context={showLogsContext.messageContext}
-        />
+      {error ? (
+        <ErrorDisplay error={error} />
+      ) : (
+        <BlueprintsPageLayout installables={installables} />
       )}
-      {(showShareContext?.extensionId || showPublishContext?.extensionId) && (
-        <ConvertToRecipeModal />
-      )}
-      {showShareContext?.blueprintId && <ShareRecipeModal />}
-      {showPublishContext?.blueprintId && <PublishRecipeModal />}
-      {body}
+      <Modals />
     </div>
   );
 };
