--- src/options/pages/blueprints/modals/ExtensionLogsModal.module.scss
@@ -1,18 +1,0 @@
-.root {
-  .modalDialog {
-    max-width: 90vw;
-
-    @media (min-width: 992px) {
-      max-width: 80vw;
-    }
-  }
-
-  .modalContent {
-    background-color: #ffffff;
-  }
-}
-
-.body {
-  max-height: 80vh;
-  overflow-y: auto;
-}
