+++ src/options/pages/blueprints/modals/Modals.tsx
@@ -1,0 +1,51 @@
+/*
+ * Copyright (C) 2022 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import React from "react";
+import { useSelector } from "react-redux";
+import { selectModalsContext } from "./blueprintModalsSelectors";
+import ConvertToRecipeModal from "./convertToRecipeModal/ConvertToRecipeModal";
+import ExtensionLogsModal from "./extensionLogsModal/ExtensionLogsModal";
+import PublishRecipeModal from "./shareModals/PublishRecipeModal";
+import ShareRecipeModal from "./shareModals/ShareRecipeModal";
+
+const Modals: React.FunctionComponent = () => {
+  const { showLogsContext, showShareContext, showPublishContext } =
+    useSelector(selectModalsContext);
+
+  const showLogsModal = showLogsContext !== null;
+  const showConvertToRecipeModal =
+    showShareContext?.extensionId || showPublishContext?.extensionId;
+  const showShareRecipeModal = showShareContext?.blueprintId;
+  const showPublishRecipeModal = showPublishContext?.blueprintId;
+
+  return (
+    <>
+      {showLogsModal && (
+        <ExtensionLogsModal
+          title={showLogsContext.title}
+          context={showLogsContext.messageContext}
+        />
+      )}
+      {showConvertToRecipeModal && <ConvertToRecipeModal />}
+      {showShareRecipeModal && <ShareRecipeModal />}
+      {showPublishRecipeModal && <PublishRecipeModal />}
+    </>
+  );
+};
+
+export default Modals;
