+++ src/options/pages/blueprints/modals/extensionLogsModal/ExtensionLogsModal.module.scss
@@ -15,9 +15,21 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-.row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 1.5rem 0;
+.root {
+  .modalDialog {
+    max-width: 90vw;
+
+    @media (min-width: 992px) {
+      max-width: 80vw;
+    }
+  }
+
+  .modalContent {
+    background-color: #ffffff;
+  }
+}
+
+.body {
+  max-height: 80vh;
+  overflow-y: auto;
 }
