+++ src/options/pages/blueprints/modals/shareModals/ShareRecipeModal.tsx
@@ -42,7 +42,7 @@ import {
 } from "@fortawesome/free-solid-svg-icons";
 import { RequireScope } from "@/auth/RequireScope";
 import ReactSelect from "react-select";
-import styles from "./ShareRecipeModal.module.scss";
+import styles from "./ShareModals.module.scss";
 import { selectAuth } from "@/auth/authSelectors";
 import { type Organization, UserRole } from "@/types/contract";
 import Loading from "./Loading";
